+++ changelog.md
@@ -5,6 +5,7 @@ Changelog
 ----
 * Course certificates can be enabled/disabled
 * Improve labelling of course privacy
+* Allow admins to download student certificates
 
 1.13.2
 ------